# Reduce memory usage by changing wildcard pattern regex to `contains` match