--- a/src/detections/configs.rs
+++ b/src/detections/configs.rs
@@ -236,7 +236,7 @@ impl StoredStatic {
                                 )
                             }
                         };
-                        ret.insert(format!("{}_{}", provider, eid), details.to_string());
+                        ret.insert(format!("{provider}_{eid}"), details.to_string());
                         Ok(())
                     })
                     .ok();