--- a/src/detections/rule/condition_parser.rs
+++ b/src/detections/rule/condition_parser.rs
@@ -129,7 +129,7 @@ impl ConditionCompiler {
 
         let result = self.compile_condition_body(&replaced_condition, name_2_node);
         if let Result::Err(msg) = result {
-            Result::Err(format!("A condition parse error has occured. {}", msg))
+            Result::Err(format!("A condition parse error has occurred. {msg}"))
         } else {
             result
         }
@@ -413,7 +413,7 @@ impl ConditionCompiler {
                 let ref_node = RefSelectionNode::new(selection_node);
                 return Result::Ok(Box::new(ref_node));
             } else {
-                let err_msg = format!("{} is not defined.", selection_name);
+                let err_msg = format!("{selection_name} is not defined.");
                 return Result::Err(err_msg);
             }
         }
@@ -1243,7 +1243,8 @@ mod tests {
         check_rule_parse_error(
             rule_str,
             vec![
-                "A condition parse error has occured. An unusable character was found.".to_string(),
+                "A condition parse error has occurred. An unusable character was found."
+                    .to_string(),
             ],
         );
     }
@@ -1266,7 +1267,7 @@ mod tests {
         check_rule_parse_error(
             rule_str,
             vec![
-                "A condition parse error has occured. ')' was expected but not found.".to_string(),
+                "A condition parse error has occurred. ')' was expected but not found.".to_string(),
             ],
         );
     }
@@ -1289,7 +1290,7 @@ mod tests {
         check_rule_parse_error(
             rule_str,
             vec![
-                "A condition parse error has occured. '(' was expected but not found.".to_string(),
+                "A condition parse error has occurred. '(' was expected but not found.".to_string(),
             ],
         );
     }
@@ -1312,7 +1313,7 @@ mod tests {
         check_rule_parse_error(
             rule_str,
             vec![
-                "A condition parse error has occured. ')' was expected but not found.".to_string(),
+                "A condition parse error has occurred. ')' was expected but not found.".to_string(),
             ],
         );
     }
@@ -1332,7 +1333,7 @@ mod tests {
         details: 'Service name : %param1%¥nMessage : Event Log Service Stopped¥nResults: Selective event log manipulation may follow this event.'
         "#;
 
-        check_rule_parse_error(rule_str,vec!["A condition parse error has occured. Unknown error. Maybe it is because there are multiple names of selection nodes.".to_string()]);
+        check_rule_parse_error(rule_str,vec!["A condition parse error has occurred. Unknown error. Maybe it is because there are multiple names of selection nodes.".to_string()]);
     }
 
     #[test]
@@ -1353,7 +1354,7 @@ mod tests {
         check_rule_parse_error(
             rule_str,
             vec![
-                "A condition parse error has occured. An illegal logical operator(and, or) was found."
+                "A condition parse error has occurred. An illegal logical operator(and, or) was found."
                     .to_string(),
             ],
         );
@@ -1377,7 +1378,7 @@ mod tests {
         check_rule_parse_error(
             rule_str,
             vec![
-                "A condition parse error has occured. An illegal logical operator(and, or) was found."
+                "A condition parse error has occurred. An illegal logical operator(and, or) was found."
                     .to_string(),
             ],
         );
@@ -1398,7 +1399,7 @@ mod tests {
         details: 'Service name : %param1%¥nMessage : Event Log Service Stopped¥nResults: Selective event log manipulation may follow this event.'
         "#;
 
-        check_rule_parse_error(rule_str,vec!["A condition parse error has occured. The use of a logical operator(and, or) was wrong.".to_string()]);
+        check_rule_parse_error(rule_str,vec!["A condition parse error has occurred. The use of a logical operator(and, or) was wrong.".to_string()]);
     }
 
     #[test]
@@ -1418,7 +1419,7 @@ mod tests {
 
         check_rule_parse_error(
             rule_str,
-            vec!["A condition parse error has occured. An illegal not was found.".to_string()],
+            vec!["A condition parse error has occurred. An illegal not was found.".to_string()],
         );
     }
 
@@ -1439,7 +1440,7 @@ mod tests {
 
         check_rule_parse_error(
             rule_str,
-            vec!["A condition parse error has occured. Not is continuous.".to_string()],
+            vec!["A condition parse error has occurred. Not is continuous.".to_string()],
         );
     }
 }