## What Changed

- Changed to use [std::string::String:contains match](https://doc.rust-lang.org/std/string/struct.String.html#method.contains) instead of regex when the following both conditions are met.
   - The pattern has two asterisks.
   - The position of the asterisk is the first and last.
     - e.g. `*abc*`(regex: `.*abc.*`)
       - Because regex which has  first and last `.*` are very slow, so replace them with [std::string::String:contains match](https://doc.rust-lang.org/std/string/struct.String.html#method.contains) + [std::string::String:to_lowercase](https://doc.rust-lang.org/std/string/struct.String.html#method.to_lowercase)
- Fixed: https://github.com/Yamato-Security/hayabusa/issues/897
  - This fix increases the number of [Benchmark2(hayabusa-sample-evtx)](Benchmark2)  detections.

## Evidence

### Environment
- OS: Windows 10 Home edition
- Hard: Memory 16GB , Core 8, SSD, laptop

### Benchmark1
I ran a benchmark using [this procedure(6.1GB evtx)](https://github.com/Yamato-Security/hayabusa/issues/778#issuecomment-1296504766) and the results were as follows.

|Version| Elapsed time | Memory(peak)  | Events with hits / Total events | Output file size(bytes)|
|-------| ------------ | --------|--------|--------| 
| [before](https://github.com/Yamato-Security/hayabusa/tree/improve-speed-by-reduce-regex-match-when-pipe)   | 00:10:42.195| 4.4 GiB | 1,593,715 / 4,817,181 | 575085385 | 
| [This PR](https://github.com/Yamato-Security/hayabusa/tree/improve-speed-by-optimize-wildcard-pattern-regex)   | 00:10:43.432 | 3.8 GiB | 1,593,715 / 4,817,181 | 575085385 | 

#### Console output

[before](https://github.com/Yamato-Security/hayabusa/tree/improve-speed-by-reduce-regex-match-when-pipe)
```
PS C:\tmp\hayabusa-2.1.0-win-64-bit> .\hayabusa.exe csv-timeline -d ..\hayabusa-big-evtx\ -o 1.csv --debug
Results Summary:

Events with hits / Total events: 1,593,715 / 4,817,181 (Data reduction: 3,223,466 events (66.92%))

Total | Unique detections: 1,627,284 | 150
Total | Unique critical detections: 0 (0.00%) | 0 (0.00%)
Total | Unique high detections: 12,044 (0.74%) | 20 (13.33%)
Total | Unique medium detections: 11,118 (0.68%) | 38 (25.33%)
Total | Unique low detections: 1,053,623 (64.75%) | 42 (28.00%)
Total | Unique informational detections: 550,499 (33.83%) | 50 (33.33%)
...

Saved file: 1.csv (575.1 MB)
Elapsed time: 00:10:42.195
Rule Parse Processing Time: 00:00:20.234
Analysis Processing Time: 00:10:00.446
Output Processing Time: 00:00:21.513

Memory usage stats:
heap stats:    peak      total      freed    current       unit      count
  reserved:    4.4 GiB    4.4 GiB   83.0 MiB    4.4 GiB
 committed:    4.0 GiB   65.0 GiB   61.2 GiB    3.8 GiB
```


[This PR](https://github.com/Yamato-Security/hayabusa/tree/improve-speed-by-optimize-wildcard-pattern-regex) 
```
PS C:\tmp\hayabusa-2.1.0-win-64-bit> .\hayabusa.exe csv-timeline -d ..\hayabusa-big-evtx\ -o 1.csv --debug
Results Summary:

Events with hits / Total events: 1,593,715 / 4,817,181 (Data reduction: 3,223,466 events (66.92%))

Total | Unique detections: 1,627,284 | 150
Total | Unique critical detections: 0 (0.00%) | 0 (0.00%)
Total | Unique high detections: 12,044 (0.74%) | 20 (13.33%)
Total | Unique medium detections: 11,118 (0.68%) | 38 (25.33%)
Total | Unique low detections: 1,053,623 (64.75%) | 42 (28.00%)
Total | Unique informational detections: 550,499 (33.83%) | 50 (33.33%)
...

Saved file: 1.csv (575.1 MB)
Elapsed time: 00:10:43.432
Rule Parse Processing Time: 00:00:19.076
Analysis Processing Time: 00:10:02.780
Output Processing Time: 00:00:21.574

Memory usage stats:
heap stats:    peak      total      freed    current       unit      count
  reserved:    3.8 GiB    3.8 GiB   83.0 MiB    3.7 GiB
 committed:    3.4 GiB   57.8 GiB   54.6 GiB    3.2 GiB
```

### Benchmark2
I ran a benchmark using [hayabusa-sample-evtx](https://github.com/Yamato-Security/hayabusa-sample-evtx) and the results were as follows.

|Version| Elapsed time | Memory(peak)  | Events with hits / Total events | Output file size(bytes)|
|-------| ------------ | --------|--------|--------|
| [before](https://github.com/Yamato-Security/hayabusa/tree/improve-speed-by-reduce-regex-match-when-pipe)   | 00:00:15.422| 1.6 GiB| 19,606 / 47,458 | 16447790 | 
| [This PR](https://github.com/Yamato-Security/hayabusa/tree/improve-speed-by-optimize-wildcard-pattern-regex)   | 00:00:13.865 |  960.1 MiB | 19,606 / 47,458 | 16450446 | 

With this fix, the `11` logs described in #897 are now detected, so the number of high/medium level detections will increase as follows.

before: `Total | Unique high detections: 6,203 (19.16%) | 274 (45.36%)`
This PR: `Total | Unique high detections: 6,213 (19.18%) | 275 (45.38%)`

before:  `Total | Unique medium detections: 1,621 (5.01%) | 184 (30.46%) `
This PR:  `Total | Unique medium detections: 1,622 (5.01%) | 185 (30.53%) `

#### Console output
[before](https://github.com/Yamato-Security/hayabusa/tree/improve-speed-by-reduce-regex-match-when-pipe)
```
Results Summary:

Events with hits / Total events: 19,606 / 47,458 (Data reduction: 27,852 events (58.69%))

Total | Unique detections: 32,375 | 604
Total | Unique critical detections: 47 (0.15%) | 19 (3.15%)
Total | Unique high detections: 6,203 (19.16%) | 274 (45.36%)
Total | Unique medium detections: 1,621 (5.01%) | 184 (30.46%)
Total | Unique low detections: 6,374 (19.69%) | 74 (12.25%)
Total | Unique informational detections: 18,130 (56.00%) | 53 (8.77%)

Dates with most total detections:
critical: 2019-07-19 (15), high: 2016-09-20 (3,643), medium: 2021-04-22 (197), low: 2016-09-20 (3,724), informational: 2016-08-19 (2,105)

Top 5 computers with most unique detections:
critical: MSEDGEWIN10 (6), IEWIN7 (3), FS03.offsec.lan (2), rootdc1.offsec.lan (2), srvdefender01.offsec.lan (2)
high: MSEDGEWIN10 (123), IEWIN7 (73), FS03.offsec.lan (32), fs03vuln.offsec.lan (31), IE10Win7 (23)
medium: MSEDGEWIN10 (68), IEWIN7 (43), fs03vuln.offsec.lan (16), FS03.offsec.lan (16), IE10Win7 (15)
low: MSEDGEWIN10 (33), FS03.offsec.lan (16), IEWIN7 (15), fs03vuln.offsec.lan (14), fs01.offsec.lan (10)
informational: IEWIN7 (18), MSEDGEWIN10 (18), fs01.offsec.lan (15), PC01.example.corp (14), FS03.offsec.lan (13)

╭───────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────╮
│ Top critical alerts:                                                       Top high alerts:                               │
├╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┤
│ Sticky Key Like Backdoor Usage - Registry (8)                              Metasploit SMB Authentication (3,562)          │
│ Active Directory Replication from Non Machine Account (6)                  Malicious Svc Possibly Installed (271)         │
│ Meterpreter or Cobalt Strike Getsystem Service Installation - System (6)   Susp Svc Installed (257)                       │
│ WannaCry Ransomware (4)                                                    Suspicious Service Installation Script (250)   │
│ Defender Alert (Severe) (4)                                                PowerShell Scripts Installed as Services (250) │
├╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┤
│ Top medium alerts:                                                         Top low alerts:                                │
├╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┤
│ Potentially Malicious PwSh (235)                                           Logon Failure (Wrong Password) (3,564)         │
│ Proc Injection (104)                                                       Susp CmdLine (Possible LOLBIN) (1,418)         │
│ Reg Key Value Set (Sysmon Alert) (103)                                     Non Interactive PowerShell (325)               │
│ Suspicious Remote Thread Target (93)                                       Rare Service Installations (321)               │
│ Wscript Execution from Non C Drive (61)                                    Proc Access (156)                              │
├╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┤
│ Top informational alerts:                                                                                                 │
├╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┤
│ Proc Exec (11,173)                                                         Explicit Logon (342)                           │
│ NetShare File Access (2,564)                                               Svc Installed (331)                            │
│ PwSh Scriptblock (789)                                                     New Non-USB PnP Device (268)                   │
│ PwSh Pipeline Exec (680)                                                   Logon (Network) (228)                          │
│ NetShare Access (433)                                                      File Created (210)                             │
╰──────────────────────────────────────────────────────────────────────────╌────────────────────────────────────────────────╯

Saved file: 1.csv (16.4 MB)
Elapsed time: 00:00:15.325
Rule Parse Processing Time: 00:00:02.315
Analysis Processing Time: 00:00:12.329
Output Processing Time: 00:00:00.679

Memory usage stats:
heap stats:    peak      total      freed    current       unit      count
  reserved:    1.6 GiB    1.6 GiB      0        1.6 GiB
 committed:    1.2 GiB    2.6 GiB    1.4 GiB    1.1 GiB
     reset:      0          0          0          0                            ok
   touched:   64.2 KiB    4.0 MiB    4.2 GiB   -4.2 GiB                        ok
  segments:     30         65         57          8                            not all freed!
-abandoned:      0          0          0          0                            ok
   -cached:      0          0          0          0                            ok
     pages:      0          0       42.3 Ki   -42.3 Ki                         ok
-abandoned:      0          0          0          0                            ok
 -extended:      0
 -noretire:      0
     mmaps:      0
   commits:    1.4 Ki
   threads:     16         16          0         16                            not all freed!
  searches:     0.0 avg
numa nodes:       1
   elapsed:      15.366 s
   process: user: 70.718 s, system: 2.078 s, faults: 374901, rss: 1.1 GiB, commit: 1.2 GiB
```


[This PR](https://github.com/Yamato-Security/hayabusa/tree/improve-speed-by-optimize-wildcard-pattern-regex) 
```
Results Summary:

Events with hits / Total events: 19,606 / 47,458 (Data reduction: 27,852 events (58.69%))

Total | Unique detections: 32,386 | 606
Total | Unique critical detections: 47 (0.15%) | 19 (3.14%)
Total | Unique high detections: 6,213 (19.18%) | 275 (45.38%)
Total | Unique medium detections: 1,622 (5.01%) | 185 (30.53%)
Total | Unique low detections: 6,374 (19.68%) | 74 (12.21%)
Total | Unique informational detections: 18,130 (55.98%) | 53 (8.75%)

Dates with most total detections:
critical: 2019-07-19 (15), high: 2016-09-20 (3,645), medium: 2021-04-22 (197), low: 2016-09-20 (3,724), informational: 2016-08-19 (2,105)

Top 5 computers with most unique detections:
critical: MSEDGEWIN10 (6), IEWIN7 (3), FS03.offsec.lan (2), rootdc1.offsec.lan (2), srvdefender01.offsec.lan (2)
high: MSEDGEWIN10 (123), IEWIN7 (74), fs03vuln.offsec.lan (32), FS03.offsec.lan (32), IE10Win7 (23)
medium: MSEDGEWIN10 (68), IEWIN7 (43), fs03vuln.offsec.lan (16), FS03.offsec.lan (16), IE10Win7 (15)
low: MSEDGEWIN10 (33), FS03.offsec.lan (16), IEWIN7 (15), fs03vuln.offsec.lan (14), fs01.offsec.lan (10)
informational: IEWIN7 (18), MSEDGEWIN10 (18), fs01.offsec.lan (15), PC01.example.corp (14), FS03.offsec.lan (13)

╭───────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────╮
│ Top critical alerts:                                                       Top high alerts:                               │
├╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┤
│ Sticky Key Like Backdoor Usage - Registry (8)                              Metasploit SMB Authentication (3,562)          │
│ Active Directory Replication from Non Machine Account (6)                  Malicious Svc Possibly Installed (271)         │
│ Meterpreter or Cobalt Strike Getsystem Service Installation - System (6)   Susp Svc Installed (257)                       │
│ WannaCry Ransomware (4)                                                    Suspicious Service Installation Script (250)   │
│ Defender Alert (Severe) (4)                                                PowerShell Scripts Installed as Services (250) │
├╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┤
│ Top medium alerts:                                                         Top low alerts:                                │
├╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┤
│ Potentially Malicious PwSh (235)                                           Logon Failure (Wrong Password) (3,564)         │
│ Proc Injection (104)                                                       Susp CmdLine (Possible LOLBIN) (1,418)         │
│ Reg Key Value Set (Sysmon Alert) (103)                                     Non Interactive PowerShell (325)               │
│ Suspicious Remote Thread Target (93)                                       Rare Service Installations (321)               │
│ Wscript Execution from Non C Drive (61)                                    Proc Access (156)                              │
├╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┤
│ Top informational alerts:                                                                                                 │
├╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┤
│ Proc Exec (11,173)                                                         Explicit Logon (342)                           │
│ NetShare File Access (2,564)                                               Svc Installed (331)                            │
│ PwSh Scriptblock (789)                                                     New Non-USB PnP Device (268)                   │
│ PwSh Pipeline Exec (680)                                                   Logon (Network) (228)                          │
│ NetShare Access (433)                                                      File Created (210)                             │
╰──────────────────────────────────────────────────────────────────────────╌────────────────────────────────────────────────╯

Saved file: 1.csv (16.5 MB)
Elapsed time: 00:00:13.865
Rule Parse Processing Time: 00:00:01.350
Analysis Processing Time: 00:00:11.891
Output Processing Time: 00:00:00.622

Memory usage stats:
heap stats:    peak      total      freed    current       unit      count
  reserved:  960.1 MiB  960.1 MiB      0      960.1 MiB
 committed:  589.6 MiB    1.8 GiB    1.3 GiB  484.7 MiB
     reset:      0          0          0          0                            ok
   touched:   64.2 KiB    1.0 MiB    3.7 GiB   -3.7 GiB                        ok
  segments:     14         17          9          8                            not all freed!
-abandoned:      0          0          0          0                            ok
   -cached:      0          0          0          0                            ok
     pages:      0          0       31.5 Ki   -31.5 Ki                         ok
-abandoned:      0          0          0          0                            ok
 -extended:      0
 -noretire:      0
     mmaps:      0
   commits:    1.2 Ki
   threads:     16         16          0         16                            not all freed!
  searches:     0.0 avg
numa nodes:       1
   elapsed:      13.905 s
   process: user: 67.531 s, system: 1.640 s, faults: 206498, rss: 542.5 MiB, commit: 594.2 MiB

```

I would appreciate it if you could review🙏